Add files via upload
Dive into the basics of Hadoop Storage File System (HDFS) with this user-friendly presentation. Learn how to create folders in HDFS easily and move files between your computer and HDFS. It's like learning the ABCs of managing data in a smart way! Perfect for beginners and those wanting to sharpen their skills in handling files on Hadoop. Check out this presentation for simple steps and clear explanations, making Hadoop file management a breeze!